requirements.txt
# Core
langgraph
langchain
langchain-community
langchain-huggingface
langsmith
python-dotenv
graphviz

# LLM integration
huggingface_hub
transformers
torch

# Tools & Retrieval
tavily-python

# Web UI
streamlit

# Testing & Quality
pytest
ruff
black

transformers>=4.38.2
accelerate>=0.27.2
torch>=2.2.0
langchain-community>=0.0.28
langchain-core>=0.1.29
bitsandbytes
auto-gptq